+++ src/Sulu/Bundle/ContactBundle/Resources/public/js/components/accounts/util/delete-dialog.js
@@ -10,7 +10,10 @@
 define(['sulucontact/model/account'], function(Account) {
     'use strict';
 
-    var templates = {
+    var constants = {
+            datagridInstanceName: 'accounts'
+        },
+        templates = {
             dialogEntityFoundTemplate: [
                 '<p><%= foundMessage %>:</p>',
                 '<% if (typeof list !== "undefined") { %>',
@@ -132,7 +135,7 @@ define(['sulucontact/model/account'], function(Account) {
                     processData: true,
                     success: function() {
                         if(!!this.recordRemove){
-                            this.sandbox.emit('husky.datagrid.record.remove', this.account.get('id'));
+                            this.sandbox.emit('husky.datagrid.' + constants.datagridInstanceName + '.record.remove', this.account.get('id'));
                         }
                         this.sandbox.emit('sulu.router.navigate', 'contacts/accounts');
                     }.bind(this)
@@ -194,7 +197,7 @@ define(['sulucontact/model/account'], function(Account) {
                         processData: true,
 
                         success: function() {
-                            this.sandbox.emit('husky.datagrid.record.remove', id);
+                            this.sandbox.emit('husky.datagrid.' + constants.datagridInstanceName + '.record.remove', id);
                         }.bind(this)
                     });
                 }.bind(this));